backend/requirements.txt
django
djangorestframework
dj-rest-auth
django-allauth
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
requests